+++ SOURCES/utils.c
@@ -6,7 +6,7 @@
 /*   By: mwilk <mwilk@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/04/05 06:44:52 by mwilk             #+#    #+#             */
-/*   Updated: 2016/06/01 16:11:54 by mwilk            ###   ########.fr       */
+/*   Updated: 2016/06/02 07:58:08 by mwilk            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,8 +20,8 @@ int			add_link(t_env *e, t_node *node, t_node *add)
 	i = 0;
 	if (node->links == NULL)
 	{
-		node->links = malloc(sizeof(t_node *) * e->nb_nodes);
-		ft_bzero(node->links, sizeof(t_node *) * e->nb_nodes);
+		node->links = malloc(sizeof(t_node *) * (size_t)e->nb_nodes);
+		ft_bzero(node->links, sizeof(t_node *) * (size_t)e->nb_nodes);
 	}
 	while (i < node->links_count && ft_strcmp(node->links[i]->name, add->name))
 		i++;
